# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateVitalReq:
    description: Request payload to capture new vital signs of a patient.
    properties:
      bloodPressureDiastolic:
        example: 80
        type: integer
      bloodPressureSystolic:
        example: 120
        type: integer
      bmi:
        example: 22.5
        type: number
      heightCm:
        example: 170
        type: number
      oxygenSaturation:
        example: 98
        maximum: 100
        type: number
      pulse:
        example: 72
        type: integer
      respiratoryRate:
        example: 18
        type: integer
      temperatureC:
        example: 36.5
        maximum: 45
        type: number
      weightKg:
        example: 65
        type: number
    type: object
...
  models.FailureResponse:
    description: Standard error response format with status and error message.
    properties:
      error:
        example: Bad Request
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  models.UpdateVitalReq:
    description: Request payload to update existing vital signs of a patient. All
      fields are optional.
    properties:
      bloodPressureDiastolic:
        example: 82
        minimum: 0
        type: integer
      bloodPressureSystolic:
        example: 122
        minimum: 0
        type: integer
      bmi:
        example: 23
        minimum: 0
        type: number
      heightCm:
        example: 172
        minimum: 0
        type: number
      oxygenSaturation:
        example: 97
        maximum: 100
        minimum: 0
        type: number
      pulse:
        example: 75
        minimum: 0
        type: integer
      respiratoryRate:
        example: 20
        minimum: 0
        type: integer
      temperatureC:
        example: 37
        maximum: 45
        minimum: 30
        type: number
      weightKg:
        example: 68
        minimum: 0
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: codervaidik@gmail.com
    name: Vaidik Bajpai
  description: Backend API for MediBridge, a medical record management system.
  title: MediBridge API
  version: "1.0"
paths:
...
  /v1/patient/{patientID}/vitals:
    delete:
      description: Deletes the vitals of a patient.
      parameters:
      - description: Patient ID
        in: path
        name: patientID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Delete patient's vitals
      tags:
      - Vitals
    post:
      consumes:
      - application/json
      description: Captures the vitals of a patient, including details like blood
        pressure, temperature, etc.
      parameters:
      - description: Patient ID
        in: path
        name: patientID
        required: true
        type: string
      - description: Vital Information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateVitalReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Capture patient's vitals
      tags:
      - Vitals
    put:
      consumes:
      - application/json
      description: Updates the vitals information of a patient.
      parameters:
      - description: Patient ID
        in: path
        name: patientID
        required: true
        type: string
      - description: Updated Vital Information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateVitalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Update patient's vitals
      tags:
      - Vitals
...
schemes:
- http
swagger: "2.0"